Add files via upload
The objective of this project is to perform comprehensive data analysis on the Indian Premier League (IPL) using Power BI, a powerful business intelligence tool. The project aims to extract, transform, and visualize IPL data to gain valuable insights and uncover patterns, trends, and performance metrics related to matches, teams, players, and venues.